add a new empty home fragment
