+++ app/src/main/res/navigation/nav.xml
@@ -4,12 +4,16 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav"
     app:startDestination="@id/homeFragment">
-
     <fragment
         android:id="@+id/homeFragment"
         android:name="com.github.libretube.ui.fragments.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
+    <fragment
+        android:id="@+id/trendsFragment"
+        android:name="com.github.libretube.ui.fragments.TrendsFragment"
+        android:label="fragment_trends"
+        tools:layout="@layout/fragment_trends" />
     <fragment
         android:id="@+id/subscriptionsFragment"
         android:name="com.github.libretube.ui.fragments.SubscriptionsFragment"
