+++ app/src/main/res/xml/shortcuts.xml
@@ -13,6 +13,16 @@
             android:targetClass="com.github.libretube.ui.activities.MainActivity" />
     </shortcut>
     <shortcut
+        android:shortcutId="trends"
+        android:enabled="true"
+        android:icon="@drawable/ic_trending"
+        android:shortcutShortLabel="@string/trends">
+        <intent
+            android:action="android.intent.action.VIEW"
+            android:targetPackage="com.github.libretube"
+            android:targetClass="com.github.libretube.ui.activities.MainActivity" />
+    </shortcut>
+    <shortcut
         android:shortcutId="subscriptions"
         android:enabled="true"
         android:icon="@drawable/ic_subscriptions"
